requirements.txt
# -------------------------------
# FASTAPI BACKEND
# -------------------------------
fastapi==0.115.0
uvicorn[standard]==0.30.6

# -------------------------------
# HTTP + ASYNC
# -------------------------------
httpx==0.27.0
aiohttp==3.9.5

# -------------------------------
# SCRAPING
# -------------------------------
beautifulsoup4==4.12.3
lxml==5.2.2

# -------------------------------
# ENV MANAGEMENT
# -------------------------------
python-dotenv==1.0.1

# -------------------------------
# AI / LLM
# -------------------------------
openai==1.40.6

# -------------------------------
# VECTOR DB (PINECONE)
# -------------------------------
pinecone-client==5.0.1

# -------------------------------
# DATA HANDLING
# -------------------------------
pydantic==2.8.2

# -------------------------------
# IMAGE PROCESSING (VISION)
# -------------------------------
Pillow==10.4.0
torch==2.3.1
transformers==4.43.3

# -------------------------------
# UTILITIES
# -------------------------------
numpy==1.26.4
tqdm==4.66.4